Queda funcional reading y binding en VRTest y desarrollada la clase con el EyeTracker HTC Vive

+++ CommonClasses/Experiments/bindingmanager.cpp
@@ -58,7 +58,7 @@ void BindingManager::init(ConfigurationManager *c){
 
     canvas->setBackgroundBrush(QBrush(Qt::gray));
 
-    qint32 WScreen, HScreen;
+    qreal WScreen, HScreen;
     WScreen = canvas->width();
     HScreen = canvas->height();
